Part 2
Tried to set up the player paddle to respond to they up and down keys as they are pressed.

I had a bit of trouble when i tried entering the code the first time, for some reason it was getting errors. However, i decided to delete the c# file and redo it and the code worked.

The Players' paddle works.