Part 3
I needed to make the Ai paddle temporarily move simulatneously with the player paddle. To do this i copied the code from player.cs and pasted it in Ai.cs and attempted to run the game.

The game failed to work because i forgot to change the class from Player to Ai.

Once the change was made and saved, both paddles moved as desired.