Part 6
Set up the game so that the ball will collide with the player paddle and bounce back. However there seems to be an error with the code because the ball sticks to the paddle and fires off when the paddle moves.

When the ball fires back at the ai paddle, it doesn't even collide, it just passes right through.