Part 15
Removed the code which allowed me to manipulate the AI_Paddle since that was for testing the collision detection.

I Successfully gave the AI_Paddle a basic form of Artificial Intelligence.
It can follow the ball perfectly, which makes it impossible to win.

Next step is to make it a bit more realistic in terms of difficulty.